sup_backend/tests/selenium/requirements-test.txt
pytest
pytest-django
selenium
webdriver-manager
